Add files via upload
Ranked-choice voting program. Checks each vote for if its for a valid candidate, records preferences. Counts votes for non-eliminated candidates. Prints the winner of the election (if there is one). Eliminates the least popular candidate by finding the minimum votes and checking which candidate matches. Checks if all non-elminated candidates are tied