Add files via upload
applies image filters (greyscale, sepia, mirror, and blur). The hardest was blur, it took replaced each pixel with the average of the 3x3 pixels surrounding the original pixel. I tried to do it all in one for loop but didn't realize that updating one pixel would throw off the calculation for the next pixel. Solved that by storing it in a placeholder and then using another set of for loops to update the new image from the placeholder