Add files via upload
Searches a raw file for signatures of a jpg (0xff, 0xd8, 0xff, 0xe*), once found, checks if its the first picture or a new picture. If its a new picture, it closes the file of the last picture and writes the data to a new file under a new name (000.jgp - XXX.jpg). Error checks # of inputs and for a valid file.