Add files via upload
Spell checker that stores a dictionary of words into an array of linked lists, based on the first letter of each word. To check spelling, it jumps to the first letter of the dictionary table, and compares the current word to each dictionary word and returns true if there's a match, meaning it is spelled correctly. At the end, the program frees each nodein the linked list by using a temporary variable to store the remainder of the list, freeing the current node, and replacing the node from the temporary variable.