Add files via upload
Harvard cs50 Lab 6: World Cup simulation. My work is on lines 18, 29, and 65 (marked # TODO). 

Reads in a CSV file containing team names and ratings, appending them to a dictionary and converting rating to an integer. Then simulates a tournament 1000 times, storing the winners in a dictionary and keeping count of their wins.