Add files via upload
Checks DNA of suspect against DNA database to find who the killer is. First loads DNA #'s of each person into a nested dictionary. Then opens a file of the suspect's DNA, checks the raw DNA for the number of repeating samples and stores it in a list for each 'STR' (# of times a particular sequence repeats, ex 'AGATC': 15). Then check each person's data in the database to see if their STR numbers match the suspect's STR numbers to determine who the killer is.